+++ moody/src/App.js
@@ -6,7 +6,6 @@ import Home from './pages/Home'
 import Register from './pages/SignIn'
 import SignUp from './pages/SignUp'
 import CalendarPage from './pages/Calendar'
-import Export from './pages/Export'
 import TextEditor from './pages/TextEditor'
 import JournalEntryInfo from './pages/JournalEntryInfo'
 import { QueryClient, QueryClientProvider } from "react-query";
@@ -26,7 +25,6 @@ function App() {
               <Route path='/signin' exact component={Register}/>
               <Route path='/signup' exact component={SignUp}/>
               <Route path='/calendar' exact component={CalendarPage}/>
-              <Route path='/export' exact component={Export}/>
               <Route path='/text' exact component={TextEditor}/>
               <Route path='/journalEntry/:id' component={JournalEntryInfo}/>
             </Switch>
