+++ moody/src/components/Calendar.jsx
@@ -9,7 +9,6 @@ import getDay from "date-fns/getDay";
 import withDragAndDrop from "react-big-calendar/lib/addons/dragAndDrop";
 import "react-big-calendar/lib/addons/dragAndDrop/styles.css";
 import "react-big-calendar/lib/css/react-big-calendar.css";
-import events from './events';
 import Link from '@material-ui/core/Link';
 
 const locales = {
@@ -34,50 +33,54 @@ const DnDCalendar = withDragAndDrop(Calendar);
 //   { start: new Date(), end: new Date(), title: "special event" }
 // ];
 
-class App extends React.Component{
-  state={
-    events
-  };
+function JournalCalendar(props) {
+  const events= [
+    {
+      id: 0,
+      title: "Journal1: Im feeling very sad",
+      start: new Date(2021, 3, 1), //the dates are wrong, they a month behind
+      end: new Date(2021, 3, 1)
+    },
+    {
+      id: 1,
+      title: "Journal2:I want to dance",
+      start: new Date(2021, 3, 15),
+      end: new Date(2021, 3, 15)
+    },
+    {
+      id: 2,
+      title: "Journal3:Im super angry with life",
+      start: new Date(2021, 3, 20),
+      end: new Date(2021, 3, 20)
+    },
+    {
+      id: 3,
+      title: "So Happy...",
+      start: new Date(new Date().setHours(new Date().getHours() - 3)),
+      end: new Date(new Date().setHours(new Date().getHours() - 3))
+    },
+    {
+      id: 3,
+      title: "Today was great",
+      start: new Date(new Date().setHours(new Date().getHours() - 3)),
+      end: new Date(new Date().setHours(new Date().getHours() - 3))
+    }
+  ];
 
-  onEventResize = (data) => {
-    const { start, end } = data;
-
-    this.setState((state) => {
-      state.events[0].start = start;
-      state.events[0].end = end;
-      return { events: state.events };
-    });
-  };
-
-  onEventDrop = (data) => {
-    console.log(data);
-
-  };
-
-  onDateClick = () => {
-
-        window.location.href='/text'
-        
-    };
-
-  render() {
     return (
       <div >
         <DnDCalendar
           selectable
           defaultView="month"
-          events={this.state.events}
+          events={events}
           localizer={localizer}
-          // onEventDrop={this.onEventDrop}
-          // onEventResize={this.onEventResize}
-          onDateClick={this.onDateClick}
-          onSelectSlot={(e)=>window.location.href='/text'}
+          onDateClick={(e) => console.log("dateclick", e)}
+          onSelectSlot={(e)=>console.log(e)}
           resizable
           style={{ height: "60vh" }}
         />
       </div>
     );
   }
-}
 
-export default App;
+export default JournalCalendar;
